Fix User model

+++ app/users/models.py
@@ -8,7 +8,7 @@ class User(db.Model, UserMixin):
     name = db.Column(db.String(60))
 
     oauth_app = db.Column(db.String(20))
-    oauth_user_id = db.Column(db.Integer)
+    oauth_user_id = db.Column(db.BigInteger)  # Facebook uses 64 bit user IDs
     oauth_token = db.Column(db.String(200))
     oauth_secret = db.Column(db.String(200))
